+++ src/sceneStructs.h
@@ -59,7 +59,7 @@ struct Mesh {
     std::vector<glm::ivec3> faceIndices; // for access vert pos
     std::vector<glm::vec3> faceNormals;
     std::vector<int> faceMatIndices;
-    std::vector<size_t> faceIndicesBVH; // sort by bvh
+    std::vector<int> faceIndicesBVH; // sort by bvh
 };
 
 struct Camera
